fix: separate tests with flags
