+++ tests/api/articles/articles-id-put.spec.ts
@@ -1,8 +1,6 @@
 import { HttpStatusCode } from '@_src_api/enums/api-status-code.enum';
-import { FeatureFlags } from '@_src_api/enums/feature-flags.enum';
 import { ArticleData } from '@_src_api/interfaces/article-data.interface';
 import { createHeaders } from '@_src_helpers_api/create-token.helper';
-import { enableFeatureFlag } from '@_src_helpers_api/feature-flags.helper';
 import { APIResponse, expect } from '@playwright/test';
 import { requestObjectTest as test } from 'src/fixtures/api/request-object.fixture';
 import { testUsers } from 'src/shared/fixtures/auth';
@@ -14,7 +12,6 @@ import { customDate } from 'test-data/shared/date.generator';
 
 test.describe('PUT articles/{id} endpoint tests', async () => {
   let newTitle: string;
-  let setHeaders: { [key: string]: string };
   const articles = `/api/articles`;
   const oldTitle = 'The difference between Scrum and Kanban';
   const newContent = 'Join KawQA z automatu';
@@ -22,10 +19,6 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
   const articleImage: string =
     'src\\test-data\\images\\Roasted_coffee_beans.jpg';
   let properArticleData: ArticleData;
-
-  test.beforeAll(async () => {
-    setHeaders = await createHeaders();
-  });
 
   test.beforeEach(async ({ articlesRequest }) => {
     newTitle =
@@ -42,10 +35,13 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
   test('Returns 200 OK status code when updating article', async ({
     articlesRequestLogged,
   }) => {
-    const existingArticleId = 1;
     //When
+    const newArticleId = await createNewArticle(
+      articlesRequestLogged,
+      properArticleData,
+    );
     const response: APIResponse = await articlesRequestLogged.put(
-      existingArticleId,
+      newArticleId,
       properArticleData,
     );
 
@@ -68,13 +64,14 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
   }) => {
     //Given
     const uniqueArticleId = await generateUniqueArticleId(articlesRequest);
+
     //When
     const response: APIResponse = await articlesRequestLogged.put(
       uniqueArticleId,
       properArticleData,
     );
-
     const responseBody = JSON.parse(await response.text());
+
     //Then
     expect(response.status()).toBe(HttpStatusCode.Created);
     expect
@@ -92,12 +89,16 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
   }) => {
     //Given
     // error: missing closing quotation mark
+    const setHeaders = await createHeaders();
     const malformedJson: string = `{"user_id": "${testUsers.regularUser.id}", "title: "${newTitle}", "body": "${newContent}", "date": "${articleDate}", "image": "src\\\\test-data\\\\images\\\\Roasted_coffee_beans.jpg"}`;
+
     //When
     const response: APIResponse = await request.put(articles, {
+      // TODO brakuje article ID
       headers: setHeaders,
       data: malformedJson,
     });
+
     //Then
     expect(response.status()).toBe(HttpStatusCode.BadRequest);
   });
@@ -122,12 +123,16 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
     articlesRequestLogged,
   }) => {
     //Given
-    const existingArticleId = 1;
+    const newArticleId = await createNewArticle(
+      articlesRequestLogged,
+      properArticleData,
+    );
     const articleDataWrongDate = properArticleData;
     articleDataWrongDate.date = '${articleDate+1}';
+
     //When
     const response: APIResponse = await articlesRequestLogged.put(
-      existingArticleId,
+      newArticleId,
       articleDataWrongDate,
     );
 
@@ -136,18 +141,25 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
   });
 
   test('Returns 422 Unprocessable Entity after updating article with empty body', async ({
+    articlesRequestLogged,
     request,
   }) => {
     //Given
-    const existingArticleId = 1;
+    const setHeaders = await createHeaders();
+    const newArticleId = await createNewArticle(
+      articlesRequestLogged,
+      properArticleData,
+    );
+
     //When
     const response: APIResponse = await request.put(
-      `${articles}/${existingArticleId}`,
+      `${articles}/${newArticleId}`,
       {
         headers: setHeaders,
         data: {},
       },
     );
+
     //Then
     expect(response.status()).toBe(HttpStatusCode.UnprocessableEntity);
   });
@@ -162,6 +174,7 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
     );
     const articleDataExistingTitle = properArticleData;
     articleDataExistingTitle.title = oldTitle;
+
     //When
     const response: APIResponse = await articlesRequestLogged.put(
       newArticleId,
@@ -180,105 +193,5 @@ test.describe('PUT articles/{id} endpoint tests', async () => {
     expect.soft(responseBody.date).toBe(articleDate);
     expect.soft(responseBody.image).toBe(articleImage);
     expect.soft(typeof responseBody.id === 'number').toBe(true);
-  });
-
-  // wykonywane równolegle te testy mogą powodować wywalanie się innych testów
-  // bo włączają globalną flagę
-  // pomysł - osobny projekt dla każdej flagi
-  // projekt puszczany osobno albo jako projekt zależny od tego projektu
-  // pomysł 2 - teardown puszczany po projekcie z niektórymi testami, teardown przywraca bazę, dopiero potem leci drugi projekt - opcja teardown w setupie projektu
-  test.describe('PUT articles/{id} endpoint tests with enabled feature_validate_article_title', async () => {
-    test.beforeAll(async ({ request }) => {
-      await enableFeatureFlag(request, FeatureFlags.ValidateArticleTitle, true);
-    });
-
-    test('Returns 200 OK status code when updating article', async ({
-      articlesRequestLogged,
-    }) => {
-      //Given
-      const existingArticleId = 1;
-
-      //When
-      const response: APIResponse = await articlesRequestLogged.put(
-        existingArticleId,
-        properArticleData,
-      );
-      // const response: APIResponse = await request.put(`${articles}/1`, {
-      //   headers: setHeaders,
-      //   data: {
-      //     user_id: testUsers.regularUser.id,
-      //     title: newTitle,
-      //     body: newContent,
-      //     date: articleDate,
-      //     image: articleImage,
-      //   },
-      // });
-      const responseBody = JSON.parse(await response.text());
-
-      //Then
-      expect(response.status()).toBe(HttpStatusCode.Ok);
-      expect
-        .soft(responseBody.user_id.toString())
-        .toEqual(testUsers.regularUser.id.toString());
-      expect.soft(responseBody.title).toBe(newTitle);
-      expect.soft(responseBody.body).toBe(newContent);
-      expect.soft(responseBody.date).toBe(articleDate);
-      expect.soft(responseBody.image).toBe(articleImage);
-      expect.soft(typeof responseBody.id === 'number').toBe(true);
-    });
-
-    test('Returns 422 status code when updating the article with the title equal to another article title', async ({
-      articlesRequestLogged,
-    }) => {
-      //Given
-      const existingArticleId = 1;
-      const articleDataExistingTitle = properArticleData;
-      articleDataExistingTitle.title = oldTitle;
-
-      //When
-      const response: APIResponse = await articlesRequestLogged.put(
-        existingArticleId,
-        articleDataExistingTitle,
-      );
-      const responseBody = JSON.parse(await response.text());
-
-      //Then
-      expect(response.status()).toBe(HttpStatusCode.UnprocessableEntity);
-      expect(responseBody.error.message).toBe('Field "title" is not unique!');
-    });
-
-    test('Returns 201 Created status code when creating article using PUT', async ({
-      articlesRequest,
-      articlesRequestLogged,
-    }) => {
-      //Given
-      const uniqueArticleId = await generateUniqueArticleId(articlesRequest);
-
-      //When
-      const response: APIResponse = await articlesRequestLogged.put(
-        uniqueArticleId,
-        properArticleData,
-      );
-      const responseBody = JSON.parse(await response.text());
-
-      //Then
-      expect(response.status()).toBe(HttpStatusCode.Created);
-      expect
-        .soft(responseBody.user_id.toString())
-        .toEqual(testUsers.regularUser.id.toString());
-      expect.soft(responseBody.title).toBe(newTitle);
-      expect.soft(responseBody.body).toBe(newContent);
-      expect.soft(responseBody.date).toBe(articleDate);
-      expect.soft(responseBody.image).toBe(articleImage);
-      expect.soft(typeof responseBody.id === 'number').toBe(true);
-    });
-
-    test.afterAll(async ({ request }) => {
-      await enableFeatureFlag(
-        request,
-        FeatureFlags.ValidateArticleTitle,
-        false,
-      );
-    });
   });
 });
